+++ source/renderer/zettlr-rendereripc.js
@@ -230,15 +230,6 @@ class ZettlrRendererIPC {
         this.send('app-quit')
         break
 
-      // Tell the main process to maximise, minimise, or close the window
-      case 'win-maximise':
-        // this.send('win-maximise')
-        break
-
-      case 'win-minimise':
-        // this.send('win-minimise')
-        break
-
       case 'win-close':
         this.send('win-close')
         break